requirements.txt
asgiref==3.3.1
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.7
django-environ==0.4.5
djangorestframework==3.12.2
drf-yasg==1.20.0
Markdown==3.3.3
pytz==2021.1
sqlparse==0.4.1
